requirements.txt
redis
Flask
psycopg2
python-dotenv